+++ chromium_src/chrome/browser/ui/autofill/chrome_autofill_client.cc
@@ -10,6 +10,7 @@
 #include "chrome/browser/profiles/profile.h"
 #include "chrome/browser/ui/autofill/payments/webauthn_dialog_controller_impl.h"
 #include "chrome/browser/ui/page_info/page_info_dialog.h"
+#include "components/autofill/core/browser/browser_autofill_manager.h"
 #include "components/optimization_guide/core/optimization_guide_features.h"
 #include "mojo/public/cpp/bindings/associated_receiver.h"
 
@@ -53,9 +54,29 @@ class BraveChromeAutofillClient : public ChromeAutofillClient {
   }
 };
 
+class BraveBrowserAutofillManager : public BrowserAutofillManager {
+ public:
+  using BrowserAutofillManager::BrowserAutofillManager;
+
+  bool IsAutofillEnabled() const override {
+    auto enabled = BrowserAutofillManager::IsAutofillEnabled();
+    if (client().GetProfileType() !=
+            profile_metrics::BrowserProfileType::kIncognito &&
+        client().GetProfileType() !=
+            profile_metrics::BrowserProfileType::kOtherOffTheRecordProfile) {
+      return enabled;
+    }
+    enabled = enabled &&
+              client().GetPrefs()->GetBoolean(kBraveAutofillPrivateWindows);
+    return enabled;
+  }
+};
+
 }  // namespace autofill
 
 #define WrapUnique WrapUnique(new autofill::BraveChromeAutofillClient(web_contents))); \
   if (0) std::unique_ptr<autofill::ChromeAutofillClient> dummy(
+#define BrowserAutofillManager BraveBrowserAutofillManager
 #include "src/chrome/browser/ui/autofill/chrome_autofill_client.cc"
+#undef BrowserAutofillManager
 #undef WrapUnique
